# timesheet-swagger.yaml
openapi: 3.0.3
info:
  title: Timesheet API
  description: |-
    This is a server for the Timesheet API - a simple API for tracking work hours of employees.
    The API is RESTful and uses JSON for serialization.

    Main features:
    - Authentication with JWT
    - Role-based access control (RBAC)
    - Upload images for user avatar
    - CRUD operations for employees
    - CRUD operations for projects
    - CRUD operations for timesheets

  termsOfService: http://swagger.io/terms/
  contact:
    email: anhtu211168@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:3000
...
paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: Register
      description: Register a new user
      operationId: register
      requestBody:
        description: User object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: Register successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
  /auth/login:
    post:
      tags:
        - auth
      summary: Login
      description: Login with username and password
      operationId: login
      requestBody:
        description: User credentials
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginBody'
        required: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthentication'

        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: User doesn't exist
  /auth/refresh-token:
    post:
      tags:
        - auth
      summary: Refresh token
      description: Refresh token
      operationId: refreshToken
      requestBody:
        description: Refresh token
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImEyNGQ3ZjI1LWMwMzktNDRiNC05NDJhLTQ3MjhiN2RmNzc1YiIsInVzZXJuYW1lIjoiYWRtaW4iLCJpYXQiOjE3MTk5MDIwNjksImV4cCI6MTcxOTk4ODQ2OX0.AwHmkN3iAlDzBpjUf08OCkFFhRlyhVw0bvLbZJCi_fM
        required: true
      responses:
        '200':
          description: Refresh token successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthentication'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
...
  /tasks:
    post:
      tags:
        - tasks
      summary: Add a new task
      description: Add a new task
      operationId: addTask
      requestBody:
        description: Task object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Create task successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
    patch:
      tags:
        - tasks
      summary: Update an existing task
      description: Update an existing task by ID
      operationId: updateTask
      requestBody:
        description: Update an existent task
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Update task successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid ID supplied
        '404':
          description: Task not found
        '401':
          description: Unauthorized
  /tasks/{id}:
    get:
      tags:
        - tasks
      summary: Get task by ID
      description: Get task by ID
      operationId: getTaskById
      parameters:
        - name: id
          in: path
          description: ID of task to return
          required: true
          schema:
            type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Get task by ID successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid ID supplied
        '404':
          description: Task not found
        '401':
          description: Unauthorized

    delete:
      tags:
        - tasks
      summary: Delete a task
      description: Soft Delete a task by ID
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          description: ID of task to delete
          required: true
          schema:
            type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Delete task successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Task deleted
        '400':
          description: Invalid ID supplied
        '404':
          description: Task not found
        '401':
          description: Unauthorized
  /tasks?search={search}&is_active={is_active}:
    get:
      tags:
        - tasks
      summary: Get all tasks
      description: Get all tasks
      operationId: getAllTasks
      parameters:
        - name: search
          in: query
          description: Search task by name
          required: false
          schema:
            type: string
          example: Task1
        - name: is_active
          in: query
          description: Filter task by status
          required: false
          schema:
            type: string
          example: ACTIVE
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Get all tasks successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
...
  /projects:
    get:
      tags:
        - projects
      summary: Get all projects
      description: Get all projects
      operationId: getAllProjects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Get all projects successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
    post:
      tags:
        - projects
      summary: Add a new project
      description: Add a new project
      operationId: addProject
      requestBody:
        description: Project object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Create project successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
    patch:
      tags:
        - projects
      summary: Update an existing project
      description: Update an existing project by ID
      operationId: updateProject
      parameters:
        - name: id
          in: path
          description: ID of project to update
          required: true
          schema:
            type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
      requestBody:
        description: Update an existent project
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Update project successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid ID supplied
        '404':
          description: Project not found
        '401':
          description: Unauthorized

    delete:
      tags:
        - projects
      summary: Delete a project
      description: Soft Delete a project by ID
      operationId: deleteProject
      parameters:
        - name: id
          in: path
          description: ID of project to delete
          required: true
          schema:
            type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Delete project successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Project deleted
        '400':
          description: Invalid ID supplied
        '404':
          description: Project not found
        '401':
          description: Unauthorized
...
components:
  schemas:
    LoginBody:
      type: object
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: 123456
    SuccessAuthentication:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImEyNGQ3ZjI1LWMwMzktNDRiNC05NDJhLTQ3MjhiN2RmNzc1YiIsInVzZXJuYW1lIjoiYWRtaW4iLCJpYXQiOjE3MTk5MDIwNjksImV4cCI6MTcxOTkwNTY2OX0.oMBKVp1HhthFkLf7LEJl-4WL5lG2wxU0kTb-6vLeutw
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImEyNGQ3ZjI1LWMwMzktNDRiNC05NDJhLTQ3MjhiN2RmNzc1YiIsInVzZXJuYW1lIjoiYWRtaW4iLCJpYXQiOjE3MTk5MDIwNjksImV4cCI6MTcxOTk4ODQ2OX0.AwHmkN3iAlDzBpjUf08OCkFFhRlyhVw0bvLbZJCi_fM
    User:
      type: object
      properties:
        id:
          type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
        username:
          type: string
          example: user1
        password:
          type: string
          example: 123456
        identify:
          type: string
          example: ID987654321
        fullname:
          type: string
          example: Nguyễn Nhân Viên
        sex:
          type: string
          example: FEMALE
        dob:
          type: string
          format: date-time
          example: '1992-02-02T00:00:00.000Z'
        email:
          type: string
          example: jane.doe@example.com
        phone_number:
          type: string
          example: '+1234567891'
        emergency_contact:
          type: string
          example: '+0987654322'
        place_of_origin:
          type: string
          example: CityC
        place_of_residence:
          type: string
          example: CityD
        current_address:
          type: string
          example: 456AnotherStCityD
        start_date:
          type: string
          format: date-time
          example: '2024-07-01T10:56:04.000Z'
        user_type:
          type: string
          example: STAFF
        role:
          type: string
          example: PM
        bankID:
          type: string
          example: BANK654321
        taxID:
          type: string
          example: TAX654321
        basic_trainer_id:
          type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
        level:
          $ref: '#/components/schemas/UserLevel'
        refresh_token:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        branch_id:
          type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
        position_id:
          type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
        is_active:
          type: string
          example: ACTIVE
        created_at:
          type: string
          format: date-time
          example: '2024-07-01T10:56:04.000Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-07-01T10:56:04.000Z'
    UserLevel:
      type: string
      enum: [I0, I1, I2, I3, F1, F2, F3, M1, M2, M3, S1, S2, S3]
      example: M1
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 8
        per_page:
          type: integer
          example: 2
        current_page:
          type: integer
          example: 4
        last_page:
          type: integer
          example: 4
        next_page:
          type: integer
          nullable: true
          example: null
        prev_page:
          type: integer
          nullable: true
          example: 3
...
    Project:
      type: object
      properties:
        id:
          type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
        name:
          type: string
          example: Project1
        project_type:
          type: string
          example: TRAINING
        start_date:
          type: string
          format: date-time
          example: '2024-07-01T10:56:04.000Z'
        end_date:
          type: string
          format: date-time
          example: '2024-07-01T10:56:04.000Z'
        is_active:
          type: string
          example: ACTIVE
        created_at:
          type: string
          format: date-time
          example: '2024-07-01T10:56:04.000Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-07-01T10:56:04.000Z'
    Task:
      type: object
      properties:
        id:
          type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
        name:
          type: string
          example: Task1
        project_id:
          type: string
          example: a24d7f25-c039-44b4-942a-4728b7df775b
        is_active:
          type: string
          example: ACTIVE
        created_at:
          type: string
          format: date-time
          example: '2024-07-01T10:56:04.000Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-07-01T10:56:04.000Z'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT